syscall: a bad pointer became a netlink send of nothing
The fake-netlink tier turned guest pointers it could not reach into
successful empty operations.

nl_sendmsg() read msg_iov and msg_iovlen out of the guest msghdr and,
when either read faulted, carried on with iov_va = iov_count = 0 --
which nl_first_iovec() then reported as "no segment", and
nl_take_request() as a request of zero bytes, successfully sent. A
kernel never gets that far: copy_msghdr_from_user runs before anything
is queued and a header it cannot read is EFAULT. The same held for the
iovec array itself, for a msghdr pointer of NULL, and for the message
body -- write(nlfd, unmapped, 64) reported 64 bytes sent, where
netlink_sendmsg copies the whole message out of the caller's memory
(memcpy_from_msg) and answers EFAULT having sent none of it.

Worse than the wrong return value: nl_take_request() was called anyway,
so the refused send left a synthesised reply behind. The read that
followed was then answered with a reply to a request that never went
out. The "never leave a request unanswered" rule the function is built
on only applies to a request that was accepted -- a send that failed
strands nobody, because a kernel leaves nothing behind either.

The receive side dropped its writeback errors: nl_maybe_recvfrom()
discarded nl_write_sockname()'s result, and nl_maybe_recvmsg() ignored
both the source-address copy and the msg_flags store, so a guest was
told a datagram had arrived at an address that was never written. The
ordinary socket paths in sys_net.c (addr_out, recvmsg_writeback) have
always answered EFAULT there, and the two tiers must be
indistinguishable to the guest.

Fixed by validating in the kernel's own order and honouring every
transfer:

  - nl_first_iovec() returns an errno: EFAULT for an array it cannot
    read (an unreadable pointer is not a message of no bytes), and the
    caller's own answer above UIO_MAXIOV -- EINVAL for writev
    (iovec_from_user), EMSGSIZE for sendmsg (copy_msghdr_from_user),
    mirroring iov_from_guest and msg_import.
  - nl_request_check() refuses a message the guest cannot back
    (rw_room over the whole range, as sendto already does) and one
    larger than the socket's send buffer, read from the stand-in's own
    SO_SNDBUF -- the same net.core.wmem_default a netlink socket takes
    it from. Without that bound a guest could name a length no buffer
    could hold and be told all of it was sent.
  - nl_maybe_recvmsg() reads the whole header up front and fails on any
    field it cannot get, then checks the source address, msg_namelen,
    msg_flags and msg_controllen writebacks. msg_controllen is now
    written at all (zero: no NETLINK_PKTINFO here), which is what the
    generic path does via its whole-struct writeback, so a caller cannot
    read its own request value back as an answer.
  - nl_write_sockname() answers a null address/length pair the way
    addr_out() does -- nothing to write back, no error -- rather than
    EINVAL, so getsockname on the two tiers agrees.
  - readv/writev on a fake fd narrow the count to unsigned at the call
    site, as the kernel's import_iovec does and as iov_from_guest
    already did, and readv refuses more than UIO_MAXIOV segments ahead
    of the receive.

tests/fixtures/netns_ack.c gains sendfault= and addrfault= rows. Both
run under the harness's existing two-tier comparison, which is what
makes them self-checking: the real-netlink tier answers out of the host
kernel, the forced AF_UNIX tier out of this code, and the assertion is
that a guest cannot tell them apart. Pre-fix the substituted tier
answered sendfault=write and addrfault=recvfrom while the real one
answered ok. The sendfault row ends by draining the socket, which is
what catches a refused send that queued a reply anyway.

+++ src/sys_file.c
@@ -794,7 +794,10 @@ SYSDEF(write) {
 
 SYSDEF(readv) {
     u64 nlret;   /* fake netlink socket: as in read */
-    if (nl_is_fd(c->m, (int)a0) && nl_maybe_readv(c, (int)a0, a1, a2, &nlret))
+    /* (unsigned)a2 for the same reason iov_from_guest takes it narrow below:
+     * the kernel's import_iovec truncates the count there too. */
+    if (nl_is_fd(c->m, (int)a0) &&
+        nl_maybe_readv(c, (int)a0, a1, (unsigned)a2, &nlret))
         return nlret;
     procfs_pre_read(c, (int)a0, -1);
     struct iovec iov[1024];
@@ -844,7 +847,8 @@ SYSDEF(readv) {
 }
 
 SYSDEF(writev) {
-    if (nl_is_fd(c->m, (int)a0)) return nl_writev(c, (int)a0, a1, a2);   /* as in write */
+    /* as in write; (unsigned)a2 as in readv above */
+    if (nl_is_fd(c->m, (int)a0)) return nl_writev(c, (int)a0, a1, (unsigned)a2);
     if (mfd_write_denied(c, (int)a0)) return (u64)(s64)-EPERM;
     struct iovec iov[1024];
     GIovec g[1024];
